requirements.txt
absl-py @ file:///C:/ci/absl-py_1615411407614/work
aiohttp @ file:///C:/ci/aiohttp_1614361133907/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet @ file:///C:/ci/chardet_1605303259695/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
Cython @ file:///C:/ci/cython_1618435342622/work
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1622042963536/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884419385/work
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
imutils==0.5.4
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///C:/ci/markdown_1614364005059/work
mediapipe==0.8.5
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854593605/work
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607367865052/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1620831380353/work
oauthlib==3.1.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy==1.4.1
six==1.16.0
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work